yellowlab_kiosk/requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.5.0
attrs==24.3.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.2.0
narwhals==1.23.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.10.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
streamlit-webrtc==0.47.9
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
